Add inventorypartnerdomain to app/site
Added the inventorypartnerdomain field in the App and Site objects. Propose adding these directly to the mainline, as opposed to an extension, because this is a core usecase. Better to mainline early than wait, because migration will be harder once there's more industry adoption.

Also did some minor repair/reformat of the <tr> tags in the table for the App fields.