[Protobuf] bid_response.ext.(igb|igs): Removed oneof and added repeated
Similar to https://github.com/InteractiveAdvertisingBureau/openrtb2.x/pull/131 , these two are marked as repeated in the specificaiton.

This gets a bit odd as they're also exclusive, which was enforced by the oneof; but oneofs can't be repeated. The exclusive constraint will need to be programatically provided, instead of provided by the protocol.