[Protobuf] Fledge igi.pbs as value instead of string
Similar to https://github.com/InteractiveAdvertisingBureau/openrtb2.x/pull/131 and https://github.com/InteractiveAdvertisingBureau/openrtb2.x/pull/132 but backwards from previous justifications.

In practice this is a json tree that's not wrapped in a string. The examples in https://github.com/InteractiveAdvertisingBureau/openrtb/blob/develop/extensions/community_extensions/Protected%20Audience%20Support.md show this as a string containing json (which is also valid but more recursive than normal). In practice this is more often passed as a json tree inlined in the document itself.