[Protobuf] The great flattening!
The great flattening! This outlines many of the nested proto messages into root objects, dramatically reducing the full names of the generated types in some languages.

Objects were flattened until they were directly under:
* BidRequest
* BidResponse
* Transparency
* NativeRequest
* NativeResponse

A few objects were not flattened:
* Ext messages - They would need to be renamed to include the outer object, and they're supposed to be "inside" the parent object.
* BidRequest.EID.UID - The object is small, and moving an object called UID feels real strange to me when that term is massively overloaded. I can outline it if poked.

SupplyChainNode was outlined because it contained the entire wrapper class name in its own name, if it was called Node it would not have been outlined.